+++ ios/PodFile
@@ -1,0 +1,6 @@
+platform :ios, '9.3'
+target 'AdVision' do
+  use_frameworks!
+  pod 'ViroReact', :path => '../node_modules/react-viro/ios/'
+  pod 'ViroKit', :path => '../node_modules/react-viro/ios/dist/ViroRenderer/'
+end
