+++ package.json
@@ -4,7 +4,7 @@
   "private": true,
   "rnpm": {
     "assets": [
-      "./assets/*"
+      "./assets/fonts"
     ]
   },
   "scripts": {
